+++ src/main/java/fabiofdez/snowyleaves/BetterSnowyLeaves.java
@@ -8,6 +8,7 @@ import org.slf4j.LoggerFactory;
 
 //? fabric {
 import fabiofdez.snowyleaves.platform.fabric.FabricPlatform;
+import org.slf4j.event.Level;
 //?} neoforge {
 /*import fabiofdez.snowyleaves.platform.neoforge.NeoforgePlatform;
  *///?} forge {
@@ -18,7 +19,7 @@ import fabiofdez.snowyleaves.platform.fabric.FabricPlatform;
 public class BetterSnowyLeaves {
 
 	public static final String MOD_ID = /*$ mod_id*/ "snowyleaves";
-	public static final String MOD_VERSION = /*$ mod_version*/ "1.3.1";
+	public static final String MOD_VERSION = /*$ mod_version*/ "1.4.0";
 	public static final String MOD_FRIENDLY_NAME = /*$ mod_name*/ "Better Snowy Leaves";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);